pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Codemni-ToolCallAgent"
version = "1.0.0"
authors = [
    {name = "CodexJitin", email = "codexjitin@gmail.com"},
]
description = "An intelligent agent that can call tools based on LLM decisions"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["ai", "agent", "llm", "tool-calling", "openai", "anthropic", "groq", "ollama", "gemini"]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.18.0"]
groq = ["groq>=0.4.0"]
google = ["google-generativeai>=0.3.0"]
ollama = ["ollama>=0.1.0"]
all = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "groq>=0.4.0",
    "google-generativeai>=0.3.0",
    "ollama>=0.1.0",
]

[project.urls]
"Homepage" = "https://github.com/Codexjitin/Codemni-ToolCallAgent"
"Bug Reports" = "https://github.com/Codexjitin/Codemni-ToolCallAgent/issues"
"Source" = "https://github.com/Codexjitin/Codemni-ToolCallAgent"
